Don't try to load system theme on wasm
Fixes #5705
