+++ app/settings.go
@@ -1,7 +1,6 @@
 package app
 
 import (
-	"bytes"
 	"os"
 	"path/filepath"
 
@@ -121,25 +120,6 @@ func (s *settings) apply() {
 func (s *settings) fileChanged() {
 	s.load()
 	s.apply()
-}
-
-func (s *settings) loadSystemTheme() fyne.Theme {
-	path := filepath.Join(app.RootConfigDir(), "theme.json")
-	data, err := fyne.LoadResourceFromPath(path)
-	if err != nil {
-		if !os.IsNotExist(err) {
-			fyne.LogError("Failed to load user theme file: "+path, err)
-		}
-		return theme.DefaultTheme()
-	}
-	if data != nil && data.Content() != nil {
-		th, err := theme.FromJSONReader(bytes.NewReader(data.Content()))
-		if err == nil {
-			return th
-		}
-		fyne.LogError("Failed to parse user theme file: "+path, err)
-	}
-	return theme.DefaultTheme()
 }
 
 func (s *settings) setupTheme() {
