+++ app/settings_mobile.go
@@ -2,6 +2,15 @@
 
 package app
 
+import (
+	"fyne.io/fyne/v2"
+	"fyne.io/fyne/v2/theme"
+)
+
+func (s *settings) loadSystemTheme() fyne.Theme {
+	return theme.DefaultTheme()
+}
+
 func (s *settings) watchSettings() {
 	// no-op on mobile
 }
