+++ app/settings_wasm.go
@@ -2,7 +2,10 @@
 
 package app
 
-// TODO: #2734
+import (
+	"fyne.io/fyne/v2"
+	"fyne.io/fyne/v2/theme"
+)
 
 func (s *settings) load() {
 	s.setupTheme()
@@ -11,6 +14,10 @@ func (s *settings) load() {
 
 func (s *settings) loadFromFile(path string) error {
 	return nil
+}
+
+func (s *settings) loadSystemTheme() fyne.Theme {
+	return theme.DefaultTheme()
 }
 
 func watchFile(path string, callback func()) {
